added assignment.py
