+++ Backend/suitable_lobby.py
@@ -1,4 +1,3 @@
-from itertools import combinations
 
 def too_many_players(lobby, o_lobby, num_players_needed):
     if lobby["num_players"] + o_lobby["num_players"] > num_players_needed:
